Fix missing latest posts on profile notes timeline
The notes tab reset items via afterNavigate's `items = []`, which raced
with the async load(). When the reset landed after load() had inserted
the newest window and advanced its cursor past it, the newest posts were
wiped and never refetched. This was triggered without any profile-to-
profile navigation: overwriteSlug()'s replaceState (npub -> NIP-05) re-
fires afterNavigate, re-running the reset mid-load.

Move the timeline state (items + load) into a Notes child component and
mount it keyed by pubkey via {#key data.pubkey}, removing the manual
reset. items lifecycle is now tied to the component mount, so no reset
can interrupt an in-flight load. Keying on pubkey (not slug) means
replaceState slug normalization no longer remounts/wipes the timeline.
